+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2023-03-07
+
+### Added 
+- `GlobalBlinkAlloc` - allocator suitable for use as `#[global_allocator]`
+  Based on `SyncBlinkAlloc`. Provides local proxies that may be `'static`
+  Unsafe `reset` method requires that all local proxies
+  and previous allocations are dropped.
+- `UnsafeGlobalAlloc` - more unsafe version of `GlobalBlinkAlloc`.
+  It must never be used concurrently. Only really usable in single-threaded
+  applications.
+  Requires `unsafe` block to initialize.
 
 ## [0.1.0] - 2023-02-27